+++ source/dpro/update_nodes_b_to_a.c
@@ -18,11 +18,11 @@ static void	set_target_b(t_stack_node *a, t_stack_node *b)
 {
 	t_stack_node	*target_node;
 	t_stack_node	*current_a;
-	long			best_match;
+	int				best_match;
 
 	while (b)
 	{
-		best_match = LONG_MAX;
+		best_match = INT_MAX;
 		current_a = a;
 		while (current_a)
 		{
@@ -33,7 +33,7 @@ static void	set_target_b(t_stack_node *a, t_stack_node *b)
 			}
 			current_a = current_a->next;
 		}
-		if (best_match == LONG_MAX)
+		if (best_match == INT_MAX)
 			b->target_node = find_min(a);
 		else
 			b->target_node = target_node;
